# docs/openapi-v1.yaml
...
openapi: 3.1.0
info:
  title: Gremio Internal API
  version: 1.0.0
  summary: Token-geschützte API für Boards und Karten.
  description: >-
    Interne REST-API von Gremio für **eigene Werkzeuge und Integrationen**.


    Alle Endpunkte verlangen einen persönlichen API-Token:


    ```

    Authorization: Bearer grm_…

    ```


    Tokens werden unter **Mein Konto → API-Tokens** (`/intern/konto`) erstellt. Ein Token ist an
    **einen Nutzer** gebunden und erbt dessen Rechte — er ist **keine** pauschale
    Administratorberechtigung.


    **Rechtestufe (`scope`)**: `write` erlaubt lesende und schreibende Zugriffe, `read` nur `GET`
    (sonst `403`).


    **Board-Beschränkung**: Ein Token kann auf bestimmte Boards begrenzt werden. Die Beschränkung
    wirkt **zusätzlich** zur Live-Zugriffsprüfung des Nutzers und kann nur einschränken, nie
    erweitern. Nicht erlaubte Boards verhalten sich wie nicht vorhanden (`404`).


    **Diese Dokumentation und die Spezifikation sind selbst anmeldepflichtig** (Gremio-Web-Session).
    Die Anmeldung dient nur dem Öffnen der Doku — für „Try it out" wird zusätzlich ein API-Token
    über **Authorize** benötigt, und die aufgerufene Route prüft ihn unabhängig davon erneut.


    Die **öffentliche** API für native Apps ist getrennt dokumentiert unter `/api/public/docs`.
  license:
    name: AGPL-3.0-only
servers:
  - url: /
    description: Diese Instanz
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/boards/{id}/cards:
    get:
      tags:
        - Karten
      operationId: listBoardCards
      summary: Karten eines Boards auflisten
      description: >-
        Karten des Boards, sortiert nach Spalten- und Kartenposition. Am Board **deaktivierte**
        Felder werden nicht ausgeliefert — die API zeigt nie mehr als die Weboberfläche.


        **Berechtigung:** Board-Zugriff.
      parameters:
        - name: id
          in: path
          required: true
          description: Board-ID.
          schema:
            type: integer
            format: int32
            minimum: 1
        - name: statusId
          in: query
          required: false
          description: Nur Karten dieser Spalte. Ungültige oder außerhalb des int32-Bereichs liegende Werte
            ergeben 400.
          schema:
            type: integer
            format: int32
            minimum: 1
        - &a5
          name: archived
          in: query
          required: false
          description: Archivierte (weggeräumte) Karten. Ohne Angabe werden **nur aktive** Karten geliefert;
            `true` liefert **nur** archivierte, `all` beide.
          schema:
            type: string
            enum:
              - "false"
              - "true"
              - all
      responses:
        "200":
          description: Karten des Boards.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardList"
              example:
                cards:
                  - id: 42
                    budgetMode: single
                    budgetRevision: 0
                    boardId: 1
                    statusId: 10
                    statusName: Eingegangen
                    position: 0
                    title: Grillabend am FB5
                    locationId: 4
                    archivedAt: null
                    createdAt: 2026-01-02T10:15:00.000Z
                    updatedAt: 2026-01-02T10:15:00.000Z
                    applicant: Max Mustermann
                    number: A_0042_2026
                    priorityId: 3
                    assigneeUserIds: &a3
                      - 7
                    deadline: 2026-02-01
                    approvedAmountCents: 25000
                    notes: null
        "400": &a4
          description: "Ungültige Eingabe: kaputtes JSON, unbekanntes Body-Feld, Schema-Verstoß, ein am Board
            **deaktiviertes** Feld, eine nicht zum Board gehörende `statusId`, eine unbekannte
            `priorityId`/`accountId`, ein Nutzer ohne Board-Zugriff oder `archived: true` (manuelles
            Archivieren gibt es nicht)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
              examples:
                schema:
                  summary: Schema-Verstoß
                  value:
                    error: Ungültige Eingabe.
                    issues:
                      - path: title
                        message: String must contain at least 1 character(s)
                fieldDisabled:
                  summary: Feld am Board deaktiviert
                  value:
                    error: Feld 'budgetTitle' ist auf diesem Board nicht aktiviert.
        "401": *a1
        "404": *a2
    post:
      tags:
        - Karten
      operationId: createCard
      summary: Karte auf einem Board anlegen
      description: >-
        Legt eine Karte an. Pflichtfeld ist `title`. Ohne `statusId` landet sie in der **ersten**
        Spalte, ohne `position` am Ende der Spalte. Ist die Board-Nummerierung aktiv, wird
        automatisch eine Antragsnummer vergeben.


        **Berechtigung:** Board-Zugriff **und** Token mit `scope=write`. Alle aktivierten
        Kartenfelder einschließlich `number`, `instructionDate` und `transferDate` dürfen
        geschrieben werden. Am Board deaktivierte Felder werden abgelehnt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreate"
            examples:
              minimal:
                summary: Nur Titel
                value:
                  title: Grillabend am FB5
              mitFeldern:
                summary: Mit weiteren Feldern
                value:
                  title: Grillabend am FB5
                  applicant: Max Mustermann
                  priorityId: 3
                  deadline: 2026-02-01
                  approvedAmountCents: 25000
      responses:
        "201":
          description: Karte angelegt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardEnvelope"
              example:
                card: &a6
                  id: 42
                  budgetMode: single
                  budgetRevision: 0
                  boardId: 1
                  statusId: 10
                  position: 0
                  title: Grillabend am FB5
                  locationId: 4
                  archivedAt: null
                  createdAt: 2026-01-02T10:15:00.000Z
                  updatedAt: 2026-01-02T10:15:00.000Z
                  applicant: Max Mustermann
                  number: A_0042_2026
                  priorityId: 3
                  assigneeUserIds: *a3
                  deadline: 2026-02-01
                  approvedAmountCents: 25000
                  notes: null
                budgetPositions: []
        "400": *a4
        "401": *a1
        "403": &a7
          description: "Schreibzugriff verweigert: Der Token hat nur Lese-Rechte (`scope=read`)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                readOnly:
                  summary: Token nur lesend
                  value:
                    error: Dieser Token hat nur Lese-Rechte (scope=read).
        "404": *a2
        "409": &a9
          description: "Leihvorgang-Board: Dessen Karten werden ausschließlich über das Inventar verwaltet,
            die API ist dort nur lesend."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Leihvorgang-Board: Karten werden über das Inventar verwaltet (API nur lesend)."
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API-Token
      description: >-
        Persönlicher API-Token aus **Mein Konto → API-Tokens**, gesendet als `Authorization: Bearer
        grm_…`.


        Der Token gilt nur im Rahmen der Rechte seines Nutzers. Die Anmeldung an dieser
        Dokumentationsseite ersetzt ihn **nicht**.
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Lesbare Fehlermeldung.
        hint:
          type: string
          description: Optionaler Hinweis zur Behebung (nur bei 401).
    ValidationError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        issues:
          type: array
          description: Feldbezogene Verstöße (nur bei Schema-Fehlern).
          items:
            type: object
            required:
              - path
              - message
            properties:
              path:
                type: string
                description: Punktnotation des Feldes, z. B. `title`.
              message:
                type: string
...
    Card:
      type: object
      description: Karte. **Optionale Felder erscheinen nur, wenn sie am Board aktiviert sind** — die
        Liste unten zeigt alle möglichen. Beträge sind Integer in **Cent**, Datumsfelder
        `YYYY-MM-DD`.
      required:
        - id
        - budgetMode
        - budgetRevision
        - boardId
        - statusId
        - position
        - title
        - locationId
        - archivedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int32
        boardId:
          type: integer
          format: int32
        statusId:
          type: integer
          format: int32
        statusName:
          type: string
        boardName:
          type: string
          description: Nur bei `GET /cards/{id}` und `GET /me/cards`.
        position:
          type: integer
          format: int32
        title:
          type: string
        locationId:
          type:
            - integer
            - "null"
          format: int32
          description: Herkunfts-Standort aus dem öffentlichen Antragsformular; null bei manuell angelegten
            Karten.
        archivedAt:
          type:
            - string
            - "null"
          format: date-time
          description: Zeitpunkt der Archivierung; null = aktiv.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        applicant:
          type:
            - string
            - "null"
        number:
          type:
            - string
            - "null"
          description: Antragsnummer.
        budgetTitle:
          type:
            - string
            - "null"
        budgetMode:
          type: string
          enum:
            - single
            - positions
          description: In positions mode accountId/budgetTitle are null and card amounts are automatically
            computed totals.
        budgetRevision:
          type: integer
          minimum: 0
          description: Optimistic revision required when replacing budgetPositions.
        priorityId:
          type:
            - integer
            - "null"
          format: int32
        accountId:
          type:
            - integer
            - "null"
          format: int32
        assigneeUserIds:
          type: array
          items:
            type: integer
            format: int32
        creatorUserId:
          type:
            - integer
            - "null"
          format: int32
        deadline:
          type:
            - string
            - "null"
          format: date
        meeting:
          type:
            - string
            - "null"
          format: date
        decisionRef:
          type:
            - string
            - "null"
        instructionDate:
          type:
            - string
            - "null"
          format: date
        transferDate:
          type:
            - string
            - "null"
          format: date
        requestedAmountCents:
          type:
            - integer
            - "null"
        approvedAmountCents:
          type:
            - integer
            - "null"
        actualAmountCents:
          type:
            - integer
            - "null"
        notes:
          type:
            - string
            - "null"
        applicantNote:
          type:
            - string
            - "null"
          description: Öffentlicher Hinweis, sichtbar auf der Statusseite des Antragstellers.
    CardEnvelope:
      type: object
      required:
        - card
        - budgetPositions
      properties:
        card:
          $ref: "#/components/schemas/Card"
        budgetPositions:
          type: array
          description: "GET/PATCH detail and POST response: stable ordered positions from the same snapshot as
            card totals/revision, only board-visible fields. Empty in single mode. List responses
            contain card totals only; request detail for allocations."
          items:
            $ref: "#/components/schemas/BudgetPosition"
    BudgetPosition:
      type: object
      additionalProperties: false
      required:
        - id
        - position
      description: Read-only projection. Amounts are nullable integer cents. Board-hidden fields are
        omitted. Strip position before writing; cardId is never exposed.
      properties:
        id:
          type: string
          format: uuid
        position:
          type: integer
          readOnly: true
        budgetTitle:
          type:
            - string
            - "null"
          maxLength: 60
        description:
          type:
            - string
            - "null"
          maxLength: 1000
        accountId:
          type: integer
          minimum: 1
          description: Required existing account for each stored position.
        requestedAmount:
          type:
            - integer
            - "null"
          minimum: 0
          maximum: 2000000000
        approvedAmount:
          type:
            - integer
            - "null"
          minimum: 0
          maximum: 2000000000
        actualAmount:
          type:
            - integer
            - "null"
          minimum: 0
          maximum: 2000000000
    CardList:
      type: object
      required:
        - cards
      properties:
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
    BudgetPositionWrite:
      type: object
      additionalProperties: false
      required:
        - id
        - budgetTitle
        - description
        - accountId
      description: "Stable, unique UUID per row; array order determines stored order. Omitted amounts
        remain unchanged (new rows: null; initial first row: previous card value). Hidden amounts
        must remain unchanged. Each amount and its total are limited to 2,000,000,000 cents."
      properties:
        id:
          type: string
          format: uuid
        budgetTitle:
          type:
            - string
            - "null"
          maxLength: 60
        description:
          type:
            - string
            - "null"
          maxLength: 1000
        accountId:
          type: integer
          format: int32
          minimum: 1
        requestedAmount:
          type:
            - integer
            - "null"
          minimum: 0
          maximum: 2000000000
        approvedAmount:
          type:
            - integer
            - "null"
          minimum: 0
          maximum: 2000000000
        actualAmount:
          type:
            - integer
            - "null"
          minimum: 0
          maximum: 2000000000
    CardCreate:
      allOf:
        - $ref: "#/components/schemas/CardWrite"
      required:
        - title
      description: Create with optional budgetPositions in one atomic transaction. budgetRevision may be
        omitted or 0 when positions are supplied. Without positions it must be omitted.
      properties:
        budgetRevision:
          const: 0
    CardWrite:
      type: object
      description: Schreibbare Kartenfelder. **Unbekannte Felder werden mit `400` abgelehnt**
        (Tippfehler-Schutz). Bei `PATCH` sind alle Felder optional; bei `POST` ist `title`
        erforderlich. `null` löscht ein Feld.
      additionalProperties: false
      properties:
        budgetPositions:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/BudgetPositionWrite"
          description: Atomic full replacement; PATCH requires the last read budgetRevision. POST supports
            positions directly. Cannot combine with top-level budgetTitle/accountId/amounts. Board
            fields budget_title and account must be enabled. Visible prefilled values may be changed
            on the first transition. A single row returns to single mode only if its description is
            empty.
        budgetRevision:
          type: integer
          minimum: 0
          maximum: 2147483647
          description: Only with budgetPositions. Required for PATCH; on POST omit or set 0. Stale revisions
            return 400 without changing the card.
        title:
          type: string
          minLength: 1
          maxLength: 200
        applicant:
          type:
            - string
            - "null"
          maxLength: 200
        budgetTitle:
          type:
            - string
            - "null"
          maxLength: 60
          description: Verknüpfungs-Schlüssel zur Finanzübersicht (max. 60 Zeichen, wie im
            Haushaltsplan-Editor).
        number:
          type:
            - string
            - "null"
          maxLength: 100
          description: Für jedes Board-Mitglied schreibbar, sofern das Feld am Board aktiviert ist.
        statusId:
          type: integer
          format: int32
          minimum: 1
          description: Zielspalte; muss zum Board der Karte gehören.
        position:
          type: integer
          format: int32
          minimum: 0
          description: 0-basierte Position in der Spalte.
        priorityId:
          type:
            - integer
            - "null"
          format: int32
          minimum: 1
        accountId:
          type:
            - integer
            - "null"
          format: int32
          minimum: 1
        assigneeUserIds:
          type: array
          maxItems: 50
          items:
            type: integer
            format: int32
            minimum: 1
          description: Ersetzt die Zuweisungen vollständig. Alle Nutzer brauchen Board-Zugriff.
        creatorUserId:
          type:
            - integer
            - "null"
          format: int32
          minimum: 1
          description: Muss Board-Zugriff haben.
        deadline:
          type:
            - string
            - "null"
          format: date
        meeting:
          type:
            - string
            - "null"
          format: date
        decisionRef:
          type:
            - string
            - "null"
          maxLength: 200
        instructionDate:
          type:
            - string
            - "null"
          format: date
          description: Für jedes Board-Mitglied schreibbar, sofern das Feld am Board aktiviert ist.
        transferDate:
          type:
            - string
            - "null"
          format: date
          description: Für jedes Board-Mitglied schreibbar, sofern das Feld am Board aktiviert ist.
        requestedAmountCents:
          type:
            - integer
            - "null"
          minimum: 0
          maximum: 2000000000
        approvedAmountCents:
          type:
            - integer
            - "null"
          minimum: 0
          maximum: 2000000000
        actualAmountCents:
          type:
            - integer
            - "null"
          minimum: 0
          maximum: 2000000000
        notes:
          type:
            - string
            - "null"
          maxLength: 20000
        applicantNote:
          type:
            - string
            - "null"
          maxLength: 20000
        archived:
          type: boolean
          description: Nur `false` (Wiederherstellen) ist zulässig; `true` wird mit `400` abgelehnt.